+++ src/main/java/org/exampleHotel/ui/gui/reservations/AddNewReservationScene.java
@@ -63,8 +63,8 @@ gridPane.setVgap(14);
             button.setOnAction(actionEvent -> {
                 LocalDate from = fromDateField.getValue();
                 LocalDate to = toDateField.getValue();
-                int guestId = guestField.getValue().getId();
-                int roomId = roomField.getValue().getId();
+                long guestId = guestField.getValue().getId();
+                long roomId = roomField.getValue().getId();
 
                try {
                    this.reservationService.createNewReservation(from, to, roomId, guestId);
